requirements.txt
aiohttp
azure-search-documents==11.6.0b9
azure-core 
azure-identity
python-dotenv==1.0.1
gunicorn
